Fixes unified logging dir again + Erfurt bootstrap
The logging component needs to be bootstrapped before Erfurt in order to
use a unified log dir configured for ontowiki for all logs. This was
already fixed by a previous commit, but a recent commit disabled
commented this code fragment out again.

Also we need to access the Erfurt store after bootstrapping Erfurt,
since we want to get notified early when connection to backend fails.