Fixes #89 RDFauthor broken caused by invalid JSON
The changes that were introduced with the testabitlity branch included,
that exit calls were removed. A lot of code in OntoWiki called
sendResponse manually (which was not a problem since exit was called).
With the replacment of exit calls with return calls, in some cases the
response was send twice, which resulted e.g. in invalid JSON.

The explicit sendResponse calls were removed, since they are not
required.

In addtition to that (in order to fix the OntoWiki test suite) the log
file creation procedure was adapted, such that it creates a new log file
if a file already exists, which can't be opened (permissions).