Enhance Makefile for usage by developers and users
- make deploy now uses submodules again (read-only by default)
- make install also uses submodules but changes URL to read-write
- add some helpscreen improvements
- vagrant target now removes libraries/Zend

How this works: (Thanks to Norman for pointing to [1]) The .gitmodules file contains
the read-only URLs for submodules (which is the default behaviour). The
"git submodule init" command will copy the settings from this file to the local
config file under .git. A special developer submodule make target re-configures
those URLs to the read-write ones. The same can be achieved the other way around by
issuing "make submodules" again.

[1] http://365git.tumblr.com/post/12581972612/set-up-a-submodule-where-only-you-have-write-access