Enhance extension test environment
- fix some existing extension tests
- adjust Makefile for extension tests
- add shell script for testing individual extensions
- fix error handling: response codes are kept and not always automatically turned
  into 500 errors anymore

Usage of the extension test execution script: Run the script when in OntoWiki root
directory and provide exactly one parameter that is the name of an existing
extension. The extension with the name provide needs to exist and have at least one
class that contains tests (e.g. under a sub-directory tests).