add min-height to <span> in resource view
In resource view, the <span> that holds the value for a predicate will
collapse if said value is empty (== ''). Since a collapsed HTML element
will have no space for mouse over events, editing an empty value with
RDFauthor is not possible via the pen icon. Setting a min-height to the
<span> will prevent this.