Add data for "Add Property" button
Add a data output that will be json encoded and read by RDFauthor to fill
"elsewhere in use" section of widget.property.js with properties provided by
templates. If no template is found or if extension is not activated an empty
output will be generated, so RDFauthor can fallback to default or check if widget was
called from button in properties.phtml.