Add title/heading to list view
The navigation module displays a title (e.g. a class name) that
is now passed to the list view in order to display a title. This helps
to remember, where the displayed list of instances is actually coming from.
The legacy method getSelectedClass() from the Model/Instances.php was not
used because it fails to provide a name for customized entries in the
navigation module.

NB: some additional testing and refining is probably needed