Reverting fix 91fc29
This commit reverts the changes made in commit 91fc29 because there
was an unwanted side-effect with multiple label widgets when creating a new
instance. The problem which was tackled with that commit (incorrect widgets
when using the template extension) should be fixed with a later commit in
the template extension.