Make comment optial (via ini)
The key "showCommentsForHeading" can be used to show/hide comments.
Additionally, it's possible to hide a comment by clicking on the heading.

It might be desirable to only display comments when clicking on the heading
and have them hidden by default. This can easily be achieved by setting
"display: none" to the #listComment id in the default.css (see line 2163).

Would be nice: Use a button that changes depending on the state (hidden or
not) and implement a "commentHiddenByDefault" for the default.ini

Even nicer: Save the state of the comment (hidden or not) across reloads.