Hotfix to repair the 'create instance' button
Since we added an option to disable the type in the resource list, the
'create instance' button can no longer determine the type of the resources by
parsing the HTML. In this case, we use the same workaround that is used for
empty resource lists. This is not ideal, the button will still not work
for resource lists you get by selecting 'View all resources'. A proper
solution would maybe be to include the type hidden even if it is not
displayed.